third_party/WebKit/LayoutTests/fast/dom/Window/replaceable.html

<!doctype html>
<script src="../../../resources/js-test.js"></script>
<script>

description('Tests that Replaceable attributes are writable');

shouldBeTrue('Object.getOwnPropertyDescriptor(window, "innerHeight").writable');

window.innerHeight = 42;
shouldBe('window.innerHeight', '42');

</script>
